Mali Valhall DDK r20p0 KMD
Provenance:
 f3deff212 (collaborate/EAC/v_r20p0)
 VX504X08X-BU-00000-r20p0-01rel0 - Android DDK

+++ mali_kbase/mali_kbase_mipe_proto.h
@@ -24,10 +24,10 @@
 #define _KBASE_MIPE_PROTO_H
 
 #define _BITFIELD_MASK_FIELD(pos, len) \
-	(((1 << len) - 1) << pos)
+	(((1u << len) - 1) << pos)
 
 #define _BITFIELD_SET_FIELD(pos, len, value) \
-	(_BITFIELD_MASK_FIELD(pos, len) & ((value) << pos))
+	(_BITFIELD_MASK_FIELD(pos, len) & (((u32) value) << pos))
 
 #define BITFIELD_SET(field_name, value) \
 	_BITFIELD_SET_FIELD(field_name ## _POS, field_name ## _LEN, value)
